refactor: consolidate class name utilities
