+++ apps/accelerate/package.json
@@ -25,7 +25,6 @@
     "@workspace/i18n": "workspace:*",
     "@workspace/sentry": "workspace:*",
     "@workspace/ui": "workspace:*",
-    "classnames": "^2.5.1",
     "framer-motion": "^12.23.28",
     "googleapis": "^146.0.0",
     "lucide-react": "^0.528.0",
@@ -34,7 +33,6 @@
     "react": "catalog:",
     "react-dom": "catalog:",
     "rtl-detect": "^1.1.2",
-    "tailwind-merge": "^3.3.1",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
