+++ apps/docs/package.json
@@ -32,7 +32,7 @@
     "ai": "^4.1.66",
     "cheerio": "^1.0.0",
     "class-variance-authority": "^0.7.1",
-    "classnames": "^2.5.1",
+    "clsx": "catalog:",
     "codehike": "^1.0.7",
     "date-fns": "^4.1.0",
     "date-fns-tz": "^3.2.0",
