+++ apps/media/package.json
@@ -35,7 +35,7 @@
     "@workspace/sentry": "workspace:*",
     "@workspace/ui": "workspace:*",
     "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
+    "clsx": "catalog:",
     "date-fns": "^3.6.0",
     "dayjs": "^1.11.21",
     "dompurify": "3.2.7",
