+++ apps/templates/package.json
@@ -20,7 +20,6 @@
     "@vercel/og": "^0.8.6",
     "@workspace/i18n": "workspace:*",
     "@workspace/ui": "workspace:*",
-    "classnames": "^2.5.1",
     "framer-motion": "^12.23.28",
     "lucide-react": "^0.528.0",
     "marked": "^16.4.2",
@@ -33,7 +32,6 @@
     "rtl-detect": "^1.1.2",
     "sanitize-html": "^2.17.5",
     "shiki": "^3.13.0",
-    "tailwind-merge": "^3.3.1",
     "tailwindcss-animate": "^1.0.7"
   },
   "devDependencies": {
