+++ apps/web/package.json
@@ -43,7 +43,7 @@
     "@workspace/ui": "workspace:*",
     "cheerio": "^1.0.0",
     "class-variance-authority": "^0.7.1",
-    "classnames": "^2.5.1",
+    "clsx": "catalog:",
     "date-fns": "^4.1.0",
     "date-fns-tz": "^3.2.0",
     "dayjs": "^1.11.21",
