+++ apps/web/src/app/components/utils.tsx
@@ -1,6 +1,6 @@
-import classNames, { ArgumentArray } from "classnames";
+import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
 
-export function cn(...inputs: ArgumentArray) {
-  return twMerge(classNames(inputs));
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
 }
