+++ apps/web/src/components/ai/AiCard.tsx
@@ -1,4 +1,4 @@
-import classNames from "classnames";
+import { clsx as classNames } from "clsx";
 import Image from "next/image";
 import { ArrowRight } from "react-feather";
