+++ apps/web/src/components/developers/sections/DevelopersResourcesSection/DevelopersResourceItem.tsx
@@ -1,4 +1,4 @@
-import classNames from "classnames";
+import { clsx as classNames } from "clsx";
 import { ArrowUpRight } from "react-feather";
 
 import { InlineLink, Link } from "@/utils/Link";
