+++ apps/web/src/components/ecdr/ECDRStats.tsx
@@ -1,5 +1,5 @@
 import { useTranslations } from "next-intl";
-import classNames from "classnames";
+import { clsx as classNames } from "clsx";
 import Link from "../../utils/Link";
 import { FormattedNumber } from "../SolFormattedMessage";
 import styles from "./ECDRStats.module.scss";
