+++ apps/web/src/components/hackathon/sections/HackathonHeroSection/HackathonHeroSection.tsx
@@ -1,4 +1,4 @@
-import classNames from "classnames";
+import { clsx as classNames } from "clsx";
 import Image from "next/legacy/image";
 import { useTranslations } from "next-intl";
 import Button from "@/components/shared/Button";
