+++ apps/web/src/components/hackathon/sections/HackathonPreviousWinnersSection/HackathonWinnerCard.tsx
@@ -1,4 +1,4 @@
-import classNames from "classnames";
+import { clsx as classNames } from "clsx";
 import Image from "next/legacy/image";
 import type { StaticImageData } from "next/image";
