+++ apps/web/src/components/playgg/PlayGGSignUpForm.tsx
@@ -1,4 +1,4 @@
-import classNames from "classnames";
+import { clsx as classNames } from "clsx";
 import { useMemo } from "react";
 import * as Yup from "yup";
