+++ apps/web/src/components/possible/PossibleStats.tsx
@@ -1,5 +1,5 @@
 import { useTranslations } from "next-intl";
-import classNames from "classnames";
+import { clsx as classNames } from "clsx";
 import styles from "./PossibleStats.module.scss";
 import Link from "../../utils/Link";
 import Image from "next/image";
