+++ packages/ui-chrome/package.json
@@ -60,7 +60,7 @@
     "@radix-ui/react-navigation-menu": "^1.2.14",
     "@radix-ui/react-visually-hidden": "^1.1.3",
     "@workspace/ui": "workspace:*",
-    "classnames": "^2.5.1",
+    "clsx": "catalog:",
     "react-feather": "^2.0.10",
     "tailwind-merge": "^3.3.1"
   }
