+++ packages/ui-chrome/src/newsletter-modal.tsx
@@ -2,15 +2,10 @@
 
 import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
-import classNames from "classnames";
-import { twMerge } from "tailwind-merge";
+import { cn } from "./classnames";
 import { useTranslations } from "next-intl";
 import { X } from "react-feather";
 import { useTheme } from "./theme-provider";
-
-function cn(...inputs: classNames.ArgumentArray) {
-  return twMerge(classNames(inputs));
-}
 
 const ITERABLE_BASE_URL =
   "https://links.iterable.com/lists/publicAddSubscriberForm?publicIdString=";
