+++ packages/ui/package.json
@@ -55,7 +55,7 @@
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-slot": "^1.2.4",
     "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
+    "clsx": "catalog:",
     "lucide-react": "^0.528.0",
     "tailwind-merge": "^3.3.1"
   }
