+++ pnpm-workspace.yaml
@@ -7,6 +7,7 @@ packages:
 verifyDepsBeforeRun: false
 
 catalog:
+  clsx: ^2.1.1
   next: 15.5.21
   next-intl: 4.13.0
   react: 19.2.6
